requirements-locked.txt
atomicwrites==1.1.5
attrs==18.1.0
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
biopython==1.71
boto==2.48.0
cvxopt==1.1.9
cycler==0.10.0
decorator==4.3.0
enum34==1.1.6
funcsigs==1.0.2
gitdb2==2.0.3
GitPython==2.1.10
ipdb==0.10.3
ipython==5.7.0
ipython-genutils==0.2.0
kiwisolver==1.0.1
matplotlib==2.2.2
more-itertools==4.2.0
numpy==1.14.4
pandas==0.17.1
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.3
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.6.1
python-dateutil==2.7.3
pytz==2018.4
scandir==1.7
scipy==1.1.0
seaborn==0.6.0
simplegeneric==0.8.1
six==1.11.0
smmap2==2.0.3
subprocess32==3.5.2
tox==2.9.1
traitlets==4.3.2
treetime==0.3.0
virtualenv==16.0.0
wcwidth==0.1.7
